+++ src/main/java/com/codeit/duckhu/domain/comment/conntroller/CommentController.java
@@ -1,4 +1,4 @@
-package com.codeit.duckhu.comments.conntroller;
+package com.codeit.duckhu.domain.comment.conntroller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
