+++ src/main/java/com/codeit/duckhu/domain/comment/dto/request/CommentUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.codeit.duckhu.comments.dto.request;
+package com.codeit.duckhu.domain.comment.dto.request;
 
 import lombok.Data;
