+++ src/main/java/com/codeit/duckhu/domain/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.codeit.duckhu.comments.repository;
+package com.codeit.duckhu.domain.comment.repository;
 
-import com.codeit.duckhu.comments.domain.Comment;
+import com.codeit.duckhu.domain.comment.domain.Comment;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
